pyproject.toml
[project]
name = "dio_blog_fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Junior010101",email = "marcondesjunior2209@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
